+++ src/model/__init__.py
@@ -1,0 +1,9 @@
+#!/bin/env python3
+# -*- coding: utf-8 -*-
+from .mtms_model import MtmsModel  # noqa: F401
+from . import connection
+
+__all__ = [
+    'MtmsModel',
+    'connection'
+]
