+++ src/model/connection/__init__.py
@@ -1,0 +1,8 @@
+#!/bin/env python3
+# -*- coding: utf-8 -*-
+from .server_connection import ServerConnection, ConnectionStatus   # noqa: F401
+
+__all__ = [
+    'ServerConnection',
+    'ConnectionStatus'
+]
