+++ src/mtms_cli.py
@@ -5,26 +5,20 @@ from __future__ import annotations
 from signal import signal, SIGINT
 import sys
 import os
-import time
 import datetime
 import logging
-from typing import Callable, Any, TypeVar, Generic, Optional
-from enum import Enum, auto
-from urllib.parse import urlparse
+import logging.config
 import asyncio
 import urwid
 import urwid.curses_display
-
-from variable import Variable, StateVariable, AutomaticStateVariable
+from variable import Variable, StateVariable, AutomaticStateVariable  # noqa: F401
+from model import MtmsModel
+from ui import MtmsUi
 
 VERSION = "0.0.1"
 
-# Global reusable components
-div = urwid.Divider()
-hr = urwid.Divider('-')
 
-
-def logger_setup():
+def setup_logging():
     """Sets up a logger instance.
     """
     logpath = os.path.join(os.path.dirname(os.path.realpath(__file__)), "../logs")
@@ -33,538 +27,33 @@ def logger_setup():
             f"Error. Log directory ({logpath}) does not exist. Please create manually first.")
         sys.exit(1)
 
-    logfile = f"{os.path.basename(__file__)}.log.{datetime.datetime.now().replace(microsecond=0).isoformat()}"
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.DEBUG)
-    logger_fh = logging.FileHandler(os.path.join(logpath, logfile))
-    logger_fh.setLevel(logging.DEBUG)
-    logger_formatter = logging.Formatter(
-        "%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-    logger_fh.setFormatter(logger_formatter)
-    logger.addHandler(logger_fh)
+    logfilename = f"{os.path.basename(__file__)}.log.{datetime.datetime.now().replace(microsecond=0).isoformat()}"
+    logfile = os.path.join(logpath, logfilename)
 
-    current_logfile = f"{os.path.basename(__file__)}.log.latest"
-    if os.path.exists(os.path.join(logpath, current_logfile)):
-        os.remove(os.path.join(logpath, current_logfile))
-    os.symlink(os.path.join(logpath, logfile),
-               os.path.join(logpath, current_logfile))
+    logging.basicConfig(
+        filename=logfile,
+        level=logging.DEBUG,
+        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
 
-    return logger
+    # Create a symlink to ".latest"
+    latest_logfilename = f"{os.path.basename(__file__)}.log.latest"
+    latest_logfile = os.path.join(logpath, latest_logfilename)
+    if os.path.exists(latest_logfile):
+        os.remove(latest_logfile)
+    os.symlink(logfile, latest_logfile)
+
+
+setup_logging()
+logger = logging.getLogger(__name__)
 
 
 def exit_handler(signal_received, frame):
     logger.info("SIGINT or Ctrl+C detected. Exiting.")
     raise urwid.ExitMainLoop()
 
-### Exceptions and Custom Classes ##############################################
-
-
-class ConnectionStatus(Enum):
-    UNDEFINED = -1
-    DISCONNECTED = 0
-    CONNECTED = 1
-    DISCONNECTING = 2
-    CONNECTING = 3
-    CANCELLING = 4
-
-### Variable and Sub-Classes ###################################################
-
-
-def validate_url(url, default_scheme='https'):
-    if not isinstance(url, str):
-        return None
-
-    if ("://" not in url):
-        url = f"{default_scheme}://{url}"
-
-    parse_result = urlparse(url)
-    if not all([parse_result.netloc, parse_result.scheme]):
-        return None
-    return f"{parse_result.scheme}://{parse_result.netloc}{parse_result.path}"
-
-#### Model #####################################################################
-
-
-class ServerConnection():
-    """Bogus server connection model.
-    """
-
-    def __init__(self,
-                 url: Optional[str] = None,
-                 on_connection_status_changed: Callable[[ServerConnection, ConnectionStatus, ConnectionStatus], None] = None
-                 ):
-        if url:
-            self._url = ServerConnection.validate_url(url)
-        else:
-            self._url = None
-
-        self._on_connection_status_changed = set()
-        if on_connection_status_changed:
-            self._on_connection_status_changed.add(on_connection_status_changed)
-
-        #self._connected = False
-        self._connection_status = ConnectionStatus.DISCONNECTED
-        self._connect_task = None
-        self._disconnect_task = None
-
-    @property
-    def url(self):
-        return self._url
-
-    @url.setter
-    def url(self, value):
-        self._url = ServerConnection.validate_url(value)
-
-    def _on_connection_status_changed_setter(self, value):
-        self._on_connection_status_changed.add(value)
-
-    on_connection_status_changed = property(None, _on_connection_status_changed_setter)
-
-    @property
-    def connection_status(self):
-        return self._connection_status
-
-    @connection_status.setter
-    def connection_status(self, value: ConnectionStatus):
-        old_value = self._connection_status
-        self._connection_status = value
-
-        for callback in self._on_connection_status_changed:
-            callback(self, value, old_value)
-
-    @property
-    def connected(self):
-        return self.connection_status in (ConnectionStatus.CONNECTED, ConnectionStatus.DISCONNECTING)
-
-    @property
-    def connecting(self):
-        return self.connection_status == ConnectionStatus.CONNECTING
-        # return self._connect_task is not None
-
-    async def connect(self) -> bool:
-        logger.info(f"Connecting to server \"{self.url}\"...")
-        self.connection_status = ConnectionStatus.CONNECTING
-
-        if self.url is None:
-            self.connection_status = ConnectionStatus.DISCONNECTED
-            logger.error("Could not initiate connection. URL is not set.")
-            raise ConnectionError("Could not initiate connection. URL is not set.")
-
-        if self._connect_task is not None:
-            logger.error("Could not initiate another connection. Connection process is already started.")
-            raise ConnectionError("Please disconnect or cancel the current connection attempt first.")
-
-        async def connect_coro():
-            # Bogus connection method
-            await asyncio.sleep(5)
-            return True
-
-        self._connect_task = asyncio.create_task(connect_coro())
-        await self._connect_task
-
-        try:
-            if self._connect_task.result():
-                logger.info("Connection success.")
-                self.connection_status = ConnectionStatus.CONNECTED
-                self._connect_task = None
-                return True
-            else:
-                logger.error("Connection failed.")
-                self.connection_status = ConnectionStatus.DISCONNECTED
-                self._connect_task = None
-                return False
-        except asyncio.CancelledError:
-            logger.info("Connection cancelled during the process.")
-            self.connection_status = ConnectionStatus.DISCONNECTED
-            self._connect_task = None
-            return False
-
-    async def disconnect(self) -> None:
-        if not self.connected:
-            return
-
-        self.connection_status = ConnectionStatus.DISCONNECTING
-
-        if self._disconnect_task is not None:
-            raise ConnectionError("Disconnect is already started.")
-
-        async def disconnect_coro():
-            if self._connect_task is not None:
-                self._connect_task.cancel()
-                await asyncio.sleep(2)
-                self._connect_task = None
-
-            await asyncio.sleep(1)
-
-        self._disconnect_task = asyncio.create_task(disconnect_coro())
-        await self._disconnect_task
-
-        self.connection_status = ConnectionStatus.DISCONNECTED
-        self._disconnect_task = None
-
-    def cancel(self) -> None:
-        self.connection_status = ConnectionStatus.CANCELLING
-        if self._connect_task is not None:
-            self._connect_task.cancel()
-            self._connect_task = None
-            self.connection_status = ConnectionStatus.DISCONNECTED
-
-        if self._disconnect_task is not None:
-            self._disconnect_task.cancel()
-            self._disconnect_task = None
-            self.connection_status = ConnectionStatus.CONNECTED
-
-    @staticmethod
-    def validate_url(url):
-        if not isinstance(url, str):
-            return None
-
-        parse_result = urlparse(url)
-        if not all([parse_result.netloc, parse_result.scheme]):
-            raise ValidationError(f"validation of url=\"{url}\" failed.")
-        return f"{parse_result.scheme}://{parse_result.netloc}{parse_result.path}"
-
-
-class MtmsModel():
-    def __init__(self):
-        self.server_url = Variable(
-            "localhost:5000",
-            validator=lambda _, url: (validate_url(url) is not None))
-
-        self.connection = ServerConnection()
-
-### UI Components ##############################################################
-
-
-class IntEditWithEnter(urwid.IntEdit):
-    signals = ["change", "postchange", "enter"]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def keypress(self, size, key):
-        key = super().keypress(size, key)
-
-        if key and key == 'enter':
-            self._emit('enter', self.value())
-            return None
-
-        return key
-
-
-class MtmsCliConnectButton(urwid.WidgetWrap):
-    BUTTON_TEXT = {
-        ConnectionStatus.UNDEFINED: '<<<error>>>',
-        ConnectionStatus.CONNECTED: 'Disconnect',
-        ConnectionStatus.DISCONNECTED: 'Connect',
-        ConnectionStatus.CONNECTING: 'Connecting...',
-        ConnectionStatus.DISCONNECTING: 'Disconnecting...',
-        ConnectionStatus.CANCELLING: 'Cancelling...',
-    }
-
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        self.connect_button = urwid.Button(self.BUTTON_TEXT[self.model.connection.connection_status])
-        self.main_widget = self.connect_button
-
-        def on_connection_status_changed(caller, value, old_value):
-            logger.debug(f"on_connection_status_changed called in \"{self}\" with caller=\"{caller}\", "
-                         f"value=\"{value}\", old_value=\"{old_value}\".")
-            self.connect_button.set_label(self.BUTTON_TEXT[value])
-
-        self.model.connection.on_connection_status_changed = on_connection_status_changed
-
-        async def on_connect_btn_pressed(button):
-            logger.debug(f"on_connect_btn_pressed called in \"{self}\" with button=\"{button}\".")
-
-            if self.model.connection.connection_status == ConnectionStatus.DISCONNECTED:
-                logger.info("Connecting to server...")
-
-                # Grab URL from component
-                url = self.model.server_url.value
-
-                # Store url to server connection
-                try:
-                    url = MtmsCliConnectButton.format_url(url)
-                    model.connection.url = url
-                    model.server_url.value = model.connection.url
-                except ValidationError as e:
-                    logger.error(f"Invalid URL=\"{url}\". Cannot connect. Error=\"{e}\".")
-                    return
-
-                # Initiate connection
-                try:
-                    result = await model.connection.connect()
-                    if result:
-                        logger.info("Connection successful.")
-                    else:
-                        logger.info("Connection failed.")
-                except ConnectionError as e:
-                    logger.error(f"Connection to server failed with error=\"{e}\".")
-
-            elif self.model.connection.connection_status == ConnectionStatus.CONNECTED:
-                logger.info("Disconnecting from server...")
-
-                try:
-                    await model.connection.disconnect()
-
-                except ConnectionError:
-                    logger.error("Disconnection request has already been started.")
-
-            elif self.model.connection.connection_status in (ConnectionStatus.CONNECTING, ConnectionStatus.DISCONNECTING):
-                logger.info("Terminating connect/disconnect process...")
-
-                self.model.connection.cancel()
-
-        urwid.connect_signal(self.main_widget, 'click', lambda b: asyncio.create_task(on_connect_btn_pressed(b)))
-
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-    @staticmethod
-    def format_url(url, default_scheme='https'):
-        if not isinstance(url, str):
-            return None
-
-        if ("://" not in url):
-            url = f"{default_scheme}://{url}"
-
-        parse_result = urlparse(url)
-        if not all([parse_result.netloc, parse_result.scheme]):
-            return None
-        return f"{parse_result.scheme}://{parse_result.netloc}{parse_result.path}"
-
-
-class MtmsCliServerURLField(urwid.WidgetWrap):
-    def __init__(self, model: MtmsModel, enabled: bool = True, label: str = "Server URL : "):
-        self.model = model
-        self._enabled = bool(enabled)
-        self._label = str(label)
-
-        self.server_url_edt = urwid.Edit(self._label, str(self.model.server_url), multiline=False)
-        self.server_url_txt = urwid.Text(f"{self._label}{self.model.server_url}")
-
-        def on_server_url_changed(caller, value, old_value):
-            logger.debug(f"on_server_url_changed called with in \"{self}\" caller=\"{caller}\", value=\"{value}\", old_value=\"{old_value}\".")
-
-            if value == old_value:
-                return
-            self.server_url_edt.set_edit_text(value)
-            self.server_url_txt.set_text(f"{self._label}{value}")
-
-        self.model.server_url.on_value_changed = on_server_url_changed
-
-        self.main_widget = urwid.WidgetPlaceholder(self.server_url_edt if self._enabled else self.server_url_txt)
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-    @property
-    def enabled(self):
-        return self._enabled
-
-    @enabled.setter
-    def enabled(self, value):
-        self._enabled = bool(value)
-        self.main_widget.original_widget = self.server_url_edt if self._enabled else self.server_url_txt
-
-    def enable(self, enable: bool = True):
-        self.enabled = enable
-
-    @property
-    def label(self):
-        return self._label
-
-    @label.setter
-    def label(self, value):
-        self._label = str(value)
-
-        self.server_url_edt.set_label(self._label)
-        self.server_url_txt.set_text(f"{self._label}{self.model.server_url}")
-
-
-class MtmsCliServerConnectWidget(urwid.WidgetWrap):
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        # Server URL
-        self.server_url_field = MtmsCliServerURLField(model=self.model, enabled=True)
-
-        def on_connection_status_changed(caller, value, old_value):
-            logger.debug(f"on_connection_status_changed called in \"{self}\" with caller=\"{caller}\", value=\"{value}\", old_value=\"{old_value}\".")
-
-            if value == ConnectionStatus.DISCONNECTED:
-                self.server_url_field.enabled = True
-            else:
-                self.server_url_field.enabled = False
-
-        self.model.connection.on_connection_status_changed = on_connection_status_changed
-
-        # Connect button
-        self.connect_btn = MtmsCliConnectButton(
-            model=self.model)
-
-        self.main_widget = urwid.Pile([
-            self.server_url_field,
-            self.connect_btn,
-        ])
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-
-class MtmsCliControlPanel(urwid.WidgetWrap):
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        # Control
-        self.control_txt = urwid.Text("Instrument control")
-
-        # Server URL
-        self.server_connect_widget = MtmsCliServerConnectWidget(self.model)
-
-        # Control panel
-        self.empty_panel = urwid.Text("No connection")
-        self.control_panel = urwid.Pile([
-            urwid.Text("Connected"),
-        ])
-        self.control_panel_placeholder = urwid.WidgetPlaceholder(self.empty_panel)
-
-        def on_connection_status_changed(caller, value, old_value):
-            logger.debug(f"on_connection_status_changed called in \"{self}\" with caller=\"{caller}\", value=\"{value}\", old_value=\"{old_value}\".")
-
-            if value == ConnectionStatus.CONNECTED:
-                self.control_panel_placeholder.original_widget = self.control_panel
-            else:
-                self.control_panel_placeholder.original_widget = self.empty_panel
-
-        self.model.connection.on_connection_status_changed = on_connection_status_changed
-
-        self.main_widget = urwid.Pile([
-            self.control_txt,
-            hr,
-            self.server_connect_widget,
-            div,
-            self.control_panel_placeholder,
-        ])
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-
-class MtmsCliConnectedWidget(urwid.WidgetWrap):
-    CONNECTION_TEXT = {
-        ConnectionStatus.CONNECTED: "yes",
-        ConnectionStatus.DISCONNECTED: "no",
-        ConnectionStatus.CONNECTING: "no (connecting...)",
-        ConnectionStatus.DISCONNECTING: "yes (disconnecting...)",
-        ConnectionStatus.CANCELLING: "unknown",
-        ConnectionStatus.UNDEFINED: "<<<error>>>",
-    }
-
-    def __init__(self, model: MtmsModel, connection_text="Connected : {status}"):
-        self.model = model
-        self.connection_text = connection_text
-
-        self.connected_txt = urwid.Text(self.connection_text.format(status=self.CONNECTION_TEXT[self.model.connection.connection_status]))
-
-        def on_connection_status_changed(caller, value, old_value):
-            logger.debug(f"on_connection_status_changed called in \"{self}\" with caller=\"{caller}\", value=\"{value}\", old_value=\"{old_value}\".")
-            self.connected_txt.set_text(self.connection_text.format(status=self.CONNECTION_TEXT[self.model.connection.connection_status]))
-
-        self.model.connection.on_connection_status_changed = on_connection_status_changed
-
-        self.main_widget = self.connected_txt
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-
-class MtmsCliStatusPanel(urwid.WidgetWrap):
-
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        # Status
-        self.status_txt = urwid.Text("Instrument status")
-        self.connected_widget = MtmsCliConnectedWidget(model=self.model)
-
-        self.main_widget = urwid.Pile([
-            self.status_txt,
-            hr,
-            self.connected_widget
-        ])
-
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-
-class MtmsCliLogWindow(urwid.WidgetWrap):
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        # Log window
-        now = datetime.datetime.now().replace(microsecond=0).isoformat()
-        self.log_txt = urwid.Text(f"Log started at {now}.")
-
-        self.main_widget = urwid.LineBox(urwid.Filler(self.log_txt, valign='bottom', top=1))
-
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-
-class MtmsCli(urwid.WidgetWrap):
-    def __init__(self, model: MtmsModel):
-        self.model = model
-
-        # Title
-        self.title_txt = urwid.Text(f"mTMS CLI version {VERSION}.")
-        self.subtitle_txt = urwid.Text("(c) 2021 Aalto University and Juuso Korhonen (MIT License)")
-
-        # Exit button
-        self.exit_btn = urwid.Button("Exit")
-
-        # Control
-        self.control_widget = MtmsCliControlPanel(model=self.model)
-
-        # Status
-        self.status_widget = MtmsCliStatusPanel(model=self.model)
-
-        # Put status on control side-by-side
-        self.system_widget = urwid.Columns([
-            self.control_widget,
-            self.status_widget,
-        ], dividechars=3, min_width=50)
-
-        # Main content arrangement
-        self.content = urwid.Pile([
-            self.system_widget,
-            div,
-            self.exit_btn
-        ])
-
-        self.header_widget = urwid.Filler(urwid.GridFlow([self.title_txt, self.subtitle_txt], 50, 3, 1, 'left'))
-        self.content_widget = urwid.LineBox(urwid.Filler(self.content, valign='top', bottom=1))
-        self.log_widget = MtmsCliLogWindow(model=self.model)
-
-        self.main_ui = urwid.Pile([
-            (1, self.header_widget),
-            self.content_widget,
-            (10, self.log_widget)
-        ], focus_item=1)
-
-        self.main_widget = urwid.WidgetPlaceholder(self.main_ui)
-
-        # Connect signals
-        urwid.connect_signal(self.exit_btn, 'click', lambda b: asyncio.create_task(self.on_exit_clicked(b)))
-
-        urwid.WidgetWrap.__init__(self, self.main_widget)
-
-    def freeze_ui(self):
-        # Freeze UI so nothing can be clicked
-        exit_screen = urwid.Filler(urwid.Text("Exiting... Please wait (Ctrl-C exits immediately).", align='center'), valign='middle')
-        self.main_widget.original_widget = exit_screen
-
-    async def on_exit_clicked(self, button):
-        logger.info("Exit button clicked.")
-        self.freeze_ui()
-
-        await self.model.connection.disconnect()
-
-        raise urwid.ExitMainLoop()
-
-
 ### Run function ###############################################################
+
 
 def run_ui():
     # Set up exit with Ctrl-Q
@@ -573,7 +62,7 @@ def run_ui():
             logger.info("Received exit signal.")
 
             try:
-                mtms_cli.freeze_ui()
+                mtms_ui.freeze_ui()
                 await mtms_model.connection.disconnect()
             except (NameError, AttributeError):
                 pass
@@ -584,12 +73,12 @@ def run_ui():
     mtms_model = MtmsModel()
 
     # Create CLI
-    mtms_cli = MtmsCli(model=mtms_model)
+    mtms_ui = MtmsUi(model=mtms_model, version=VERSION)
 
     # Start loop
     asyncio_loop = asyncio.get_event_loop()
     event_loop = urwid.AsyncioEventLoop(loop=asyncio_loop)
-    loop = urwid.MainLoop(mtms_cli,
+    loop = urwid.MainLoop(mtms_ui,
                           event_loop=event_loop,
                           screen=urwid.raw_display.Screen(),
                           unhandled_input=lambda key: asyncio.create_task(exit_on_q(key)))
@@ -597,8 +86,6 @@ def run_ui():
 
 
 if __name__ == "__main__":
-    global logger
-    logger = logger_setup()   # Side effect: exposes `logger`
     logger.info("Starting main program.")
     signal(SIGINT, exit_handler)   # Attach an exit listener
     run_ui()
