+++ src/ui/__init__.py
@@ -1,5 +1,7 @@
 #!/bin/env python3
 # -*- coding: utf-8 -*-
+from .mtms_ui import MtmsUi
 
 __all__ = [
+    'MtmsUi'
 ]
