+++ src/ui/helpers.py
@@ -1,0 +1,7 @@
+#!/bin/env python3
+# -*- coding: utf-8 -*-
+import urwid
+
+# Global reusable components
+div = urwid.Divider()
+hr = urwid.Divider('-')
